Match most of ftPe_Float
