+++ src/melee/ft/chara/ftPeach/ftPe_Float.h
@@ -8,7 +8,7 @@
 
 #include <placeholder.h>
 
-/* 11BA20 */ /* static */ bool ftPe_Float_CheckContinueInput(Fighter* fp);
+/* 11BA20 */ bool ftPe_Float_CheckContinueInput(Fighter* fp);
 /* 11BA54 */ bool ftPe_8011BA54(HSD_GObj* gobj);
 /* 11BAD8 */ bool ftPe_8011BAD8(HSD_GObj* gobj);
 /* 11BB6C */ void ftPe_8011BB6C(HSD_GObj* gobj, bool arg1);
