+++ src/melee/ft/chara/ftPeach/ftPe_FloatFall.c
@@ -1,5 +1,6 @@
 #include <platform.h>
-#include "ft/forward.h"
+#include <melee/ft/chara/ftPeach/forward.h>
+#include <melee/ft/forward.h>
 
 #include "ftPe_FloatFall.h"
 
@@ -9,6 +10,26 @@
 #include "ftCommon/ftCo_FallSpecial.h"
 
 #include <baselib/gobj.h>
+
+static ftPeach_MotionState getFloatDir(HSD_GObj* gobj)
+{
+    Fighter* fp = GET_FIGHTER(gobj);
+    return fp->input.lstick.x * fp->facing_dir > -p_ftCommonData->x78
+               ? ftPe_MS_FloatFallF
+               : ftPe_MS_FloatFallB;
+}
+
+void updateFloatDir(HSD_GObj* gobj)
+{
+    Fighter* fp = GET_FIGHTER(gobj);
+    ftPe_DatAttrs* da = fp->dat_attrs;
+    ftPeach_MotionState msid = getFloatDir(gobj);
+    float anim_start = msid == ftPe_MS_FloatFallF ? da->floatfallf_anim_start
+                                                  : da->floatfallb_anim_start;
+    Fighter_ChangeMotionState(gobj, msid, Ft_MF_KeepGfx,
+                              anim_start - da->floatfall_anim_start_offset,
+                              1.0f, 0.0f, NULL);
+}
 
 void ftPe_FloatFall_Anim(HSD_GObj* gobj)
 {
