+++ src/melee/ft/chara/ftPeach/ftPe_FloatFall.h
@@ -5,6 +5,7 @@
 
 #include <placeholder.h>
 
+void updateFloatDir(HSD_GObj* gobj);
 /* 11BDF0 */ void ftPe_FloatFall_Anim(HSD_GObj* gobj);
 /* 11BE2C */ void ftPe_FloatFall_IASA(HSD_GObj* gobj);
 /* 11BE30 */ void ftPe_FloatFall_Phys(HSD_GObj* gobj);
